# openapi.yaml
openapi: 3.1.0
info:
  title: Job Search Dashboard API
  version: 0.1.0
  description: REST API for the Job Search + Resume Management Dashboard
servers:
  - url: http://localhost:3000
    description: Local development
paths:
...
  /api/resumes/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      operationId: getResume
      summary: Get resume with all nested data
      tags: [Resumes]
      responses:
        '200':
          description: Full resume
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ResumeDetail'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      operationId: updateResume
      summary: Update resume
      tags: [Resumes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateResume'
      responses:
        '200':
          description: Updated resume
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Resume'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteResume
      summary: Delete resume
      tags: [Resumes]
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: string
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/resumes/{id}/positions:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      operationId: listPositions
      summary: List positions for a resume
      tags: [Positions]
      responses:
        '200':
          description: Positions with bullets
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PositionWithBullets'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      operationId: createPosition
      summary: Add position to resume
      tags: [Positions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PositionInput'
      responses:
        '201':
          description: Created position
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Position'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/resumes/{id}/positions/{positionId}/bullets/{bulletId}:
    parameters:
      - $ref: '#/components/parameters/Id'
      - $ref: '#/components/parameters/PositionId'
      - $ref: '#/components/parameters/BulletId'
    put:
      operationId: updateBullet
      summary: Update bullet
      tags: [Bullets]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBulletInput'
      responses:
        '200':
          description: Updated bullet
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Bullet'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteBullet
      summary: Delete bullet
      tags: [Bullets]
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: string
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/resumes/{id}/skills/{skillId}:
    parameters:
      - $ref: '#/components/parameters/Id'
      - $ref: '#/components/parameters/SkillId'
    put:
      operationId: updateSkillCategory
      summary: Update skill category
      tags: [Skills]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSkillCategoryInput'
      responses:
        '200':
          description: Updated skill category
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/SkillCategory'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteSkillCategory
      summary: Delete skill category
      tags: [Skills]
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: string
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/resumes/{id}/education/{educationId}:
    parameters:
      - $ref: '#/components/parameters/Id'
      - $ref: '#/components/parameters/EducationId'
    put:
      operationId: updateEducation
      summary: Update education entry
      tags: [Education]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEducationInput'
      responses:
        '200':
          description: Updated education
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Education'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteEducation
      summary: Delete education entry
      tags: [Education]
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: string
        '404':
          $ref: '#/components/responses/NotFound'
  /api/resumes/{id}/projects:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      operationId: listProjects
      summary: List projects
      tags: [Projects]
      responses:
        '200':
          description: List of projects
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Project'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      operationId: createProject
      summary: Add project
      tags: [Projects]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectInput'
      responses:
        '201':
          description: Created project
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
    PositionId:
      name: positionId
      in: path
      required: true
      schema:
        type: string
    BulletId:
      name: bulletId
      in: path
      required: true
      schema:
        type: string
    SkillId:
      name: skillId
      in: path
      required: true
      schema:
        type: string
    EducationId:
      name: educationId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              details:
                type: object
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
  schemas:
...
    ResumeType:
      type: string
      enum: [BASE, JOB_APPLICATION_OPTIMIZED, JOB_TITLE_OPTIMIZED]
...
    Resume:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        title:
          type: string
        type:
          $ref: '#/components/schemas/ResumeType'
        templateId:
          type: string
        fontFamily:
          type: string
        fontSize:
          type: string
        lineHeight:
          type: string
        sectionTitleCasing:
          type: string
        dateFormat:
          type: string
        marginH:
          type: string
        marginV:
          type: string
        pageSize:
          type: string
        sectionOrder:
          type: string
        score:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ResumeDetail:
      allOf:
        - $ref: '#/components/schemas/Resume'
        - type: object
          properties:
            profile:
              $ref: '#/components/schemas/Profile'
              nullable: true
            positions:
              type: array
              items:
                $ref: '#/components/schemas/PositionWithBullets'
            skills:
              type: array
              items:
                $ref: '#/components/schemas/SkillCategory'
            education:
              type: array
              items:
                $ref: '#/components/schemas/Education'
            projects:
              type: array
              items:
                $ref: '#/components/schemas/Project'
...
    UpdateResume:
      type: object
      properties:
        title:
          type: string
        type:
          $ref: '#/components/schemas/ResumeType'
        templateId:
          type: string
        fontFamily:
          type: string
        fontSize:
          type: string
        lineHeight:
          type: string
        sectionTitleCasing:
          type: string
        dateFormat:
          type: string
        marginH:
          type: string
        marginV:
          type: string
        pageSize:
          type: string
        sectionOrder:
          type: string
        score:
          type: integer
          nullable: true
    Profile:
      type: object
      properties:
        id:
          type: string
        resumeId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        linkedin:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        targetTitle:
          type: string
          nullable: true
        summary:
          type: string
          nullable: true
...
    Position:
      type: object
      properties:
        id:
          type: string
        resumeId:
          type: string
        company:
          type: string
        companyDesc:
          type: string
          nullable: true
        title:
          type: string
        location:
          type: string
          nullable: true
        startDate:
          type: string
          nullable: true
        endDate:
          type: string
          nullable: true
        current:
          type: boolean
        hidden:
          type: boolean
        sortOrder:
          type: integer
        sourcePositionId:
          type: string
          nullable: true
    PositionWithBullets:
      allOf:
        - $ref: '#/components/schemas/Position'
        - type: object
          properties:
            bullets:
              type: array
              items:
                $ref: '#/components/schemas/Bullet'
    PositionInput:
      type: object
      required: [company, title, sortOrder]
      properties:
        company:
          type: string
          minLength: 1
        companyDesc:
          type: string
          nullable: true
        title:
          type: string
          minLength: 1
        location:
          type: string
          nullable: true
        startDate:
          type: string
          nullable: true
        endDate:
          type: string
          nullable: true
        current:
          type: boolean
          default: false
        hidden:
          type: boolean
          default: false
        sortOrder:
          type: integer
        sourcePositionId:
          type: string
          nullable: true
...
    Bullet:
      type: object
      properties:
        id:
          type: string
        positionId:
          type: string
        content:
          type: string
        hidden:
          type: boolean
        sortOrder:
          type: integer
        sourceBulletId:
          type: string
          nullable: true
...
    UpdateBulletInput:
      type: object
      properties:
        content:
          type: string
        hidden:
          type: boolean
        sortOrder:
          type: integer
        sourceBulletId:
          type: string
          nullable: true
    SkillCategory:
      type: object
      properties:
        id:
          type: string
        resumeId:
          type: string
        name:
          type: string
        skills:
          type: string
        sortOrder:
          type: integer
...
    UpdateSkillCategoryInput:
      type: object
      properties:
        name:
          type: string
        skills:
          type: string
        sortOrder:
          type: integer
    Education:
      type: object
      properties:
        id:
          type: string
        resumeId:
          type: string
        institution:
          type: string
        degree:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        startDate:
          type: string
          nullable: true
        endDate:
          type: string
          nullable: true
        current:
          type: boolean
        achievements:
          type: string
          nullable: true
...
    UpdateEducationInput:
      type: object
      properties:
        institution:
          type: string
        degree:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        startDate:
          type: string
          nullable: true
        endDate:
          type: string
          nullable: true
        current:
          type: boolean
        achievements:
          type: string
          nullable: true
    Project:
      type: object
      properties:
        id:
          type: string
        resumeId:
          type: string
        name:
          type: string
        link:
          type: string
          nullable: true
        startDate:
          type: string
          nullable: true
        endDate:
          type: string
          nullable: true
        current:
          type: boolean
        achievements:
          type: string
          nullable: true
    ProjectInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        link:
          type: string
          nullable: true
        startDate:
          type: string
          nullable: true
        endDate:
          type: string
          nullable: true
        current:
          type: boolean
          default: false
        achievements:
          type: string
          nullable: true
...